+++ internal/pullzoneshieldresourcevalidator/shield.go
@@ -1,0 +1,8 @@
+package pullzoneshieldresourcevalidator
+
+var PlanTypeMap = map[uint8]string{
+	0: "Basic",
+	1: "Advanced",
+	2: "Business",
+	3: "Enterprise",
+}