+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 ### Added
+- resource pullzone_shield: support `Business` and `Enterprise` tiers;
 - resource pullzone: validate fields for ComputeScript origin.type ([#43](https://github.com/BunnyWay/terraform-provider-bunnynet/issues/43));
 ### Fixed
 - resource pullzone: validation for `safehop_connection_timeout` and `safehop_response_timeout`;